+++ src/internal/database/engine/mysql/mysql.go
@@ -576,7 +576,7 @@ func NewDumper(host string, port int, user, pass string, tlsCfg ...*config.TLSCo
 		Triggers:          true,
 		Views:             true,
 		connCfg: &config.ConnectivityConfig{
-			Enabled:       true,
+			Enabled:       config.BoolPtr(true),
 			Method:        config.MethodFull,
 			RetryInterval: 5,
 			Timeout:       300,